Default build type is "Release" (blank before)
This build type is noted to be used as the default in README.md, which should be reflected by the default settings.
The build type specifies the optimization (Release passes compiler flag "-O3", which improves performance heavily).
This code is similar to code used by (among others) Jan Behrens, cliutils on gitlab and david-m-rosen on github.